refactor: rename config folder to ruleset
SL-2295
