+++ src/ruleset/__tests__/__fixtures__/ruleset-invalid.json
