+++ src/ruleset/path.ts
